+++ src/s_files.h
@@ -35,17 +35,17 @@
 
 char *Slash(char *Path, int Add);
 char *SlashDir(char *Path);
-int ExpandPath(const char *Path, char *Expanded, int ExpandSize);
+int ExpandPath(const char *Path, char *Expanded, size_t ExpandSize);
 int CompletePath(const char *Path, char *Completed, int Num);
 int IsSameFile(const char *Path1, const char *Path2);
-int JustDirectory(const char *Path, char *Dir, int DirSize);
-int JustFileName(const char *Path, char *Name, int NameSize);
-int JustLastDirectory(const char *Path, char *Dir, int DirSize);
-int JustRoot(const char *Path, char *Root, int RootSize);
+int JustDirectory(const char *Path, char *Dir, size_t DirSize);
+int JustFileName(const char *Path, char *Name, size_t NameSize);
+int JustLastDirectory(const char *Path, char *Dir, size_t DirSize);
+int JustRoot(const char *Path, char *Root, size_t RootSize);
 int FileExists(const char *Path);
 int IsFullPath(const char *Path);
 int IsDirectory(const char *Path);
-const char *ShortFName(const char *Path, int len);
+const char *ShortFName(const char *Path, size_t len);
 int ChangeDir(char *Dir);
 int JoinDirFile(char *Dest, const char *Dir, const char *Name);
 char *SepRChr(char *Dir);
